+++ src/main/java/oripa/gui/presenter/creasepattern/byvalue/AngleMeasuringAction.java
@@ -1,8 +1,12 @@
-package oripa.gui.presenter.creasepattern;
+package oripa.gui.presenter.creasepattern.byvalue;
 
 import oripa.domain.paint.PaintContext;
 import oripa.domain.paint.byvalue.SelectingVertexForAngle;
 import oripa.domain.paint.byvalue.ValueSetting;
+import oripa.gui.presenter.creasepattern.AbstractGraphicMouseAction;
+import oripa.gui.presenter.creasepattern.CreasePatternViewContext;
+import oripa.gui.presenter.creasepattern.GraphicMouseAction;
+import oripa.gui.presenter.creasepattern.ObjectGraphicDrawer;
 
 public class AngleMeasuringAction extends AbstractGraphicMouseAction {
