+++ src/main/java/oripa/gui/presenter/creasepattern/copypaste/ChangeOriginAction.java
@@ -1,4 +1,4 @@
-package oripa.gui.presenter.creasepattern;
+package oripa.gui.presenter.creasepattern.copypaste;
 
 import java.util.Collection;
 
@@ -7,6 +7,10 @@ import javax.vecmath.Vector2d;
 import oripa.domain.paint.PaintContext;
 import oripa.domain.paint.copypaste.SelectionOriginHolder;
 import oripa.domain.paint.geometry.NearestItemFinder;
+import oripa.gui.presenter.creasepattern.AbstractGraphicMouseAction;
+import oripa.gui.presenter.creasepattern.CreasePatternViewContext;
+import oripa.gui.presenter.creasepattern.GraphicMouseAction;
+import oripa.gui.presenter.creasepattern.ObjectGraphicDrawer;
 import oripa.value.OriLine;
 
 public class ChangeOriginAction extends AbstractGraphicMouseAction {
