+++ src/main/java/oripa/gui/presenter/creasepattern/copypaste/CopyAndPasteAction.java
@@ -1,9 +1,14 @@
-package oripa.gui.presenter.creasepattern;
+package oripa.gui.presenter.creasepattern.copypaste;
 
 import javax.vecmath.Vector2d;
 
 import oripa.domain.paint.PaintContext;
 import oripa.domain.paint.copypaste.SelectionOriginHolder;
+import oripa.gui.presenter.creasepattern.AbstractGraphicMouseAction;
+import oripa.gui.presenter.creasepattern.CreasePatternViewContext;
+import oripa.gui.presenter.creasepattern.EditMode;
+import oripa.gui.presenter.creasepattern.GraphicMouseAction;
+import oripa.gui.presenter.creasepattern.ObjectGraphicDrawer;
 
 public class CopyAndPasteAction extends AbstractGraphicMouseAction {
