+++ src/main/java/oripa/gui/view/main/UIPanel.java
@@ -57,12 +57,12 @@ import oripa.gui.bind.binder.BinderInterface;
 import oripa.gui.bind.binder.ViewChangeBinder;
 import oripa.gui.bind.state.PaintBoundStateFactory;
 import oripa.gui.bind.state.action.PaintActionSetterFactory;
-import oripa.gui.presenter.creasepattern.AngleMeasuringAction;
 import oripa.gui.presenter.creasepattern.CreasePatternViewContext;
 import oripa.gui.presenter.creasepattern.EditMode;
-import oripa.gui.presenter.creasepattern.LengthMeasuringAction;
 import oripa.gui.presenter.creasepattern.MouseActionHolder;
+import oripa.gui.presenter.creasepattern.byvalue.AngleMeasuringAction;
 import oripa.gui.presenter.creasepattern.byvalue.AngleValueInputListener;
+import oripa.gui.presenter.creasepattern.byvalue.LengthMeasuringAction;
 import oripa.gui.presenter.creasepattern.byvalue.LengthValueInputListener;
 import oripa.gui.view.estimation.EstimationResultFrameFactory;
 import oripa.gui.view.foldability.FoldabilityCheckFrameFactory;
